+++ dyatel/dyatel_sel/core/core_checkbox.py
@@ -5,10 +5,11 @@ from logging import info
 
 from selenium.webdriver.remote.webelement import WebElement as SeleniumWebElement
 
+from dyatel.dyatel_sel.core.core_element import CoreElement
 from dyatel.internal_utils import WAIT_EL
 
 
-class CoreCheckbox:
+class CoreCheckbox(CoreElement):
 
     def __init__(self, locator: str, locator_type='', name='', parent=None, wait=False, by_attr=False):
         """
@@ -21,8 +22,7 @@ class CoreCheckbox:
         :param wait: add element waiting in `wait_page_loaded` function of CorePage
         :param by_attr: get is_checked state by custom attribute
         """
-        from dyatel.dyatel_sel.elements.web_element import CoreElement
-        self.wrapped_element = CoreElement(locator=locator, locator_type=locator_type, name=name, parent=parent, wait=wait)
+        super().__init__(locator=locator, locator_type=locator_type, name=name, parent=parent, wait=wait)
         self.checked = None
         self.by_attr = by_attr
 
@@ -33,7 +33,8 @@ class CoreCheckbox:
 
         :return: selenium WebElement
         """
-        return self.wrapped_element.element
+        self.wait_availability()
+        return self._get_element(wait=False)
 
     def is_checked(self) -> bool:
         """
@@ -51,7 +52,7 @@ class CoreCheckbox:
         :return: self
         """
         if not self.is_checked():
-            self._wait_clickable(silent=True).element.click()
+            self.wait_clickable(silent=True).element.click()
             self.checked = True
 
         return self
@@ -63,7 +64,7 @@ class CoreCheckbox:
         :return: self
         """
         if self.is_checked():
-            self._wait_clickable(silent=True).element.click()
+            self.wait_clickable(silent=True).element.click()
             self.checked = False
 
         return self
@@ -77,7 +78,7 @@ class CoreCheckbox:
         """
         return self.element.text if self.element.text else self.element.get_attribute('value')
 
-    def _wait_clickable(self, silent=False, timeout=WAIT_EL) -> CoreCheckbox:
+    def wait_clickable(self, silent=False, timeout=WAIT_EL) -> CoreCheckbox:
         """
         Wait until checkbox become clickable
         Selenium `element_to_be_clickable` doesn't work for some checkboxes
@@ -87,13 +88,34 @@ class CoreCheckbox:
         :return: self
         """
         if not silent:
-            info(f'Wait until "{self.wrapped_element.name}" become clickable')
+            info(f'Wait until "{self.name}" become clickable')
 
         start_time = time.time()
         while time.time() - start_time < timeout and not self.element.is_enabled():
             pass
 
         if not self.element.is_enabled():
-            raise Exception(f'"{self.wrapped_element.name}" not clickable')
+            raise Exception(f'"{self.name}" not clickable')
+
+        return self
+
+    def wait_availability(self, timeout=WAIT_EL, silent=False) -> CoreElement:
+        """
+        Wait for current element available in DOM
+
+        :param: timeout: time to stop waiting
+        :param: silent: erase log
+        :return: self
+        """
+        if not silent:
+            info(f'Wait until presence of "{self.name}"')
+
+        start_time = time.time()
+
+        while time.time() - start_time < timeout and not self.is_available():
+            pass
+
+        if not self.is_available():
+            raise Exception(f'Can\'t wait element in DOM "{self.name}". {self.get_element_logging_data()}')
 
         return self