+++ tests/web_tests/test_driver.py
@@ -33,12 +33,40 @@ def test_driver_execute_script_with_args(driver_wrapper, mouse_event_page):
     'appium', 'safari',  # FIXME: Unskip for playwright safari
     reason='Safari browser/Mobile platforms doesnt support multiple drivers'
 )
-def test_second_driver(driver_wrapper, second_driver_wrapper):
-    second_driver_wrapper.get(MouseEventPage().url)
-    driver_wrapper.get(PizzaOrderPage().url)
+def test_second_driver_different_page(driver_wrapper, second_driver_wrapper):
+    mouse_page = MouseEventPage().set_driver(second_driver_wrapper)
+    pizza_page = PizzaOrderPage().set_driver(driver_wrapper)
 
-    assert MouseEventPage().set_driver(second_driver_wrapper).is_page_opened()
-    assert MouseEventPage().header_logo.is_displayed()
+    assert pizza_page.driver != mouse_page.driver
+    assert pizza_page.driver_wrapper != mouse_page.driver_wrapper
+    assert mouse_page.header_logo.driver != pizza_page.quantity_input.driver
+    assert mouse_page.header_logo.driver_wrapper != pizza_page.quantity_input.driver_wrapper
 
-    assert PizzaOrderPage().set_driver(driver_wrapper).is_page_opened()
-    assert PizzaOrderPage().quantity_input.is_displayed()
+    mouse_page.open_page()
+    pizza_page.open_page()
+
+    assert mouse_page.is_page_opened()
+    assert mouse_page.header_logo.is_displayed()
+
+    assert pizza_page.is_page_opened()
+    assert pizza_page.quantity_input.is_displayed()
+
+
+@pytest.mark.skip_platform(
+    'appium', 'safari',  # FIXME: Unskip for playwright safari
+    reason='Safari browser/Mobile platforms doesnt support multiple drivers'
+)
+def test_second_driver_same_page(driver_wrapper, second_driver_wrapper):
+    mouse_page1 = MouseEventPage().set_driver(second_driver_wrapper)
+    mouse_page2 = MouseEventPage().set_driver(driver_wrapper)
+
+    assert mouse_page1.driver != mouse_page2.driver
+    assert mouse_page1.driver_wrapper != mouse_page2.driver_wrapper
+    assert mouse_page1.header_logo.driver != mouse_page2.header_logo.driver
+    assert mouse_page1.header_logo.driver_wrapper != mouse_page2.header_logo.driver_wrapper
+
+    mouse_page1.open_page()
+    mouse_page2.open_page()
+
+    assert mouse_page1.is_page_opened()
+    assert mouse_page2.is_page_opened()